# spiffworkflow-backend/src/spiffworkflow_backend/api.yml
openapi: "3.0.2"
info:
  version: 1.0.0
  title: spiffworkflow-backend
  license:
    name: LGPL
servers:
  - url: http://localhost:5000/v1.0
# this is handled in flask now
security: []
paths:
...
  /public/messages/form/{modified_message_name}:
    parameters:
      - name: modified_message_name
        in: path
        required: true
        description: The message_name, modified to replace slashes (/) with colons
        schema:
          type: string
    get:
      tags:
        - Messages
      operationId: spiffworkflow_backend.routes.public_controller.message_form_show
      summary: Gets the form associated with the given message name.
      responses:
        "200":
          description: The json schema form.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
  /public/messages/submit/{modified_message_name}:
    parameters:
      - name: modified_message_name
        in: path
        required: true
        description: The message_name, modified to replace slashes (/) with colons
        schema:
          type: string
      - name: execution_mode
        in: query
        required: false
        description: Either run in "synchronous" or "asynchronous" mode.
        schema:
          type: string
          enum:
            - synchronous
            - asynchronous
    post:
      tags:
        - Messages
      operationId: spiffworkflow_backend.routes.public_controller.message_form_submit
      summary: Instantiate and run a given process model with a message start event matching given name
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                properties:
                  task_data:
                    $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
                  process_instance:
                    $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
  /public/tasks/{process_instance_id}/{task_guid}:
    parameters:
      - name: task_guid
        in: path
        required: true
        description: The unique id of an existing process group.
        schema:
          type: string
      - name: process_instance_id
        in: path
        required: true
        description: The unique id of an existing process instance.
        schema:
          type: integer
      - name: execution_mode
        in: query
        required: false
        description: Either run in "synchronous" or "asynchronous" mode.
        schema:
          type: string
          enum:
            - synchronous
            - asynchronous
    get:
      tags:
        - Tasks
      operationId: spiffworkflow_backend.routes.public_controller.form_show
      summary: Gets one task that a user wants to complete
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
    put:
      tags:
        - Tasks
      operationId: spiffworkflow_backend.routes.public_controller.form_submit
      summary: Update the form data for a tasks
      requestBody:
        required: false
        content:
          application/json:
            schema:
              nullable: true
              allOf:
                - $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
        "202":
          description: "ok: true"
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OkTrue"
  /public/tasks/{process_instance_id}/{task_guid}/send-user-signal-event:
    parameters:
      - name: process_instance_id
        in: path
        required: true
        description: The unique id of an existing process instance.
        schema:
          type: integer
      - name: task_guid
        in: path
        required: true
        description: The unique id of an existing task.
        schema:
          type: string
    post:
      tags:
        - Tasks
      operationId: spiffworkflow_backend.routes.public_controller.signal_submit
      summary: Send a signal event to a process instance from a guest form
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AwesomeUnspecifiedPayload"
      responses:
        "200":
          description: One task
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Task"
...
components:
  schemas:
...
    OkTrue:
      properties:
        ok:
          type: boolean
          example: true
...
    Task:
      properties:
        id:
          readOnly: true
          type: string
          description: Task identifier (same as task_guid for consistency)
        task_id:
          type: string
          description: Task identifier
        task_guid:
          type: string
          description: Globally unique identifier for this task instance
        task_name:
          type: string
          description: BPMN element name/ID
        name:
          type: string
          description: Human-readable task name
        task_title:
          type: string
          description: Display title for the task
        title:
          type: string
          description: Display title (alias for task_title)
        task_type:
          type: string
          description: Type of BPMN task (UserTask, ScriptTask, etc.)
          example: "UserTask"
        type:
          type: string
          description: Type of task
        task_status:
          type: string
          description: Current status of the task
          example: "READY"
        state:
          type: string
          description: Task state
        completed:
          type: boolean
          description: Whether the task has been completed
        documentation:
          type: string
          description: Markdown documentation for the task
        data:
          type: object
          description: Task data/variables
        form_file_name:
          type: string
          nullable: true
          description: JSON schema file name for the form
        ui_form_file_name:
          type: string
          nullable: true
          description: UI schema file name for the form
        json_metadata:
          type: object
          description: Additional metadata stored as JSON
        multi_instance_type:
          type: string
          enum: ["none", "looping", "parallel", "sequential"]
        multi_instance_count:
          type: number
        multi_instance_index:
          type: number
        process_name:
          type: string
        process_instance_id:
          type: integer
          description: ID of the process instance this task belongs to
        process_model_identifier:
          type: string
          description: Identifier of the process model
        process_model_display_name:
          type: string
          description: Display name of the process model
        process_instance_status:
          type: string
          description: Current status of the process instance
        bpmn_process_identifier:
          type: string
          description: BPMN process definition identifier
        last_milestone_bpmn_name:
          type: string
          nullable: true
          description: Name of the last milestone reached
        potential_owner_usernames:
          type: string
          description: Comma-separated list of potential owner usernames
        process_initiator_username:
          type: string
          description: Username of the user who started the process
        lane_name:
          type: string
          nullable: true
          description: BPMN lane name if task is in a lane
        lane_assignment_id:
          type: integer
          nullable: true
          description: ID of the lane assignment
        actual_owner_id:
          type: integer
          nullable: true
          description: User ID of the actual owner
        completed_by_user_id:
          type: integer
          nullable: true
          description: User ID of who completed the task
        assigned_user_group_identifier:
          type: string
          nullable: true
          description: Identifier of the assigned user group
        created_at_in_seconds:
          type: integer
          description: Unix timestamp when task was created
        updated_at_in_seconds:
          type: integer
          description: Unix timestamp when task was last updated
        process_instance_summary:
          type: string
          nullable: true
          description: Summary of the process instance
        properties:
          type: object
          description: Additional task properties
      example:
        id: study_identification
        name: Study Identification
        title: IRB Review
        documentation: "# Heading 1


          Markdown documentation text goes here"
        type: form
        state: ready
        form:
          "key": "irb_review_form"
          "fields":
            - "id": "irb_review_type"
              "type": "enum"
              "label": "Select IRB Review Type"
              "options":
                - id: "emergency_use"
                  name: "Emergency Use"
                - id: "humanitarian_device"
                  name: "Humanitarian Device"
                - id: "non_human"
                  name: "Non-Human"
                - id: "non_uva_agent"
                  name: "Non-UVA Agent"
                - id: "exempt"
                  name: "Exempt"
                - id: "non_engaged"
                  name: "Non-Engaged"
                - id: "expedited"
                  name: "Expedited"
                - id: "full_board"
                  name: "Full Board"
              "default_value": "Full Board"
              "validation":
                - name: "required"
                  config: "true"
              "properties":
                - id: "description"
                  value: "Description text goes here"
                - id: "help"
                  value: "# Heading 1


                    Markdown help text goes here"
                - id: "required_expression"
                  value: "model.my_boolean_field_id && model.my_enum_field_value !== 'something'"
                - id: "hide_expression"
                  value: "model.my_enum_field_value === 'something'"
...
    AwesomeUnspecifiedPayload:
      type: "object"
      additionalProperties: {}
...